Organizing based on functionality instead of file type
