+++ src/app/game/components/guess-list/guess-list.component.scss
