+++ src/app/game/components/guess-list/guess-list.component.ts
@@ -1,0 +1,22 @@
+import { CommonModule } from '@angular/common';
+import { Component, inject } from '@angular/core';
+import { GuessService } from '../../services/guess/guess.service';
+import { GuessType } from '../../models/guess-type';
+
+@Component({
+  selector: 'app-guess-list',
+  standalone: true,
+  imports: [
+    CommonModule
+  ],
+  templateUrl: './guess-list.component.html',
+  styleUrls: ['./guess-list.component.scss']
+})
+export class GuessListComponent {
+
+  guessType = GuessType;
+
+  private guessService = inject(GuessService);
+  guesses$ = this.guessService.guesses$;
+
+}
