+++ src/app/game/components/progress-bar/progress-bar.component.scss
