+++ src/app/game/models/constants.ts
