+++ src/app/game/models/game-constants.ts
