+++ src/app/game/models/guess-type.ts
