+++ src/app/game/services/progress-bar/progress-bar.service.spec.ts
