+++ src/app/game/state/models/game-state.model.ts
