+++ src/app/game/state/reducers/global-state.reducer.ts
@@ -1,0 +1,14 @@
+import { createReducer, on } from "@ngrx/store";
+import { resetGameState, setGameState } from "../actions/game-state.actions";
+import { GameState } from "../models/game-state.model";
+
+export const initialState: GameState = {
+  guessCount: 0,
+  isPlayingMusic: false
+}
+
+export const GlobalReducer = createReducer<GameState>(
+  initialState,
+  on(setGameState, (state: GameState) => state),
+  on(resetGameState, () => initialState)
+)
