+++ src/app/game/state/selectors/game-state.selector.ts
@@ -1,0 +1,9 @@
+import { createFeatureSelector, createSelector } from "@ngrx/store";
+import { GameState } from "../models/game-state.model";
+
+export const getGlobalState = createFeatureSelector<GameState>('gameState');
+
+export const selectGameState = createSelector(
+  getGlobalState,
+  (state: GameState) => (state)
+);
