--- src/app/services/spotify/album/album.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { AlbumService } from './album.service';
-
-describe('AlbumService', () => {
-  let service: AlbumService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(AlbumService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
