--- src/app/services/spotify/artist/artist.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ArtistService } from './artist.service';
-
-describe('ArtistService', () => {
-  let service: ArtistService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ArtistService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
