--- src/app/services/spotify/artist/artist.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable({
-  providedIn: 'root'
-})
-export class ArtistService {
-
-  constructor() { }
-}
