--- src/app/services/spotify/playlist/playlist.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { PlaylistService } from './playlist.service';
-
-describe('PlaylistService', () => {
-  let service: PlaylistService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(PlaylistService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
