--- src/app/services/spotify/song/song.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { SongService } from './song.service';
-
-describe('SongService', () => {
-  let service: SongService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(SongService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
