--- src/app/state/models/global-state.model.ts
@@ -1,5 +1,0 @@
-import { GameState } from "./game-state.model";
-
-export interface GlobalState {
-  gameState: GameState;
-}
