--- src/app/state/selectors/game-state.selector.ts
@@ -1,9 +1,0 @@
-import { createFeatureSelector, createSelector } from "@ngrx/store";
-import { GlobalState } from "../models/global-state.model";
-
-export const getGlobalState = createFeatureSelector<GlobalState>('globalState');
-
-export const selectGameState = createSelector(
-  getGlobalState,
-  (state: GlobalState) => (state.gameState)
-);
