+++ src/app/task-scheduler/models/woozle-task-state.ts
